# docs/swagger.yaml
definitions:
...
  main.CreateSeatPayload:
    properties:
      number:
        minimum: 1
        type: integer
      room_id:
        minimum: 1
        type: integer
      row:
        minimum: 1
        type: integer
    required:
    - number
    - room_id
    - row
    type: object
  main.CreateTicketPayload:
    properties:
      price:
        minimum: 0
        type: number
      seat_id:
        minimum: 1
        type: integer
      session_id:
        minimum: 1
        type: integer
      user_id:
        minimum: 1
        type: integer
    required:
    - price
    - seat_id
    - session_id
    type: object
...
  main.UpdateSeatPayload:
    properties:
      row:
        minimum: 1
        type: integer
      seat_number:
        minimum: 1
        type: integer
    type: object
...
  main.UpdateTicketPayload:
    properties:
      price:
        minimum: 0
        type: number
      user_id:
        minimum: 1
        type: integer
    type: object
...
  store.Movie:
    properties:
      created_at:
        type: string
      description:
        type: string
      duration:
        type: integer
      id:
        type: integer
      poster_url:
        type: string
      release_date:
        type: string
      title:
        type: string
    type: object
...
  store.Room:
    properties:
      capacity:
        type: integer
      id:
        type: integer
      name:
        type: string
    type: object
  store.Seat:
    properties:
      id:
        type: integer
      room_id:
        type: integer
      row:
        type: integer
      seat_number:
        type: integer
    type: object
  store.SeatWithMetadata:
    properties:
      id:
        type: integer
      price:
        type: number
      room_id:
        type: integer
      row:
        type: integer
      seat_number:
        type: integer
      status:
        type: string
    type: object
  store.Session:
    properties:
      id:
        type: integer
      movie:
        $ref: '#/definitions/store.Movie'
      movie_id:
        type: integer
      price:
        type: number
      room:
        $ref: '#/definitions/store.Room'
      room_id:
        type: integer
      start_time:
        type: string
    type: object
...
  store.Ticket:
    properties:
      created_at:
        type: string
      id:
        type: integer
      price:
        type: number
      seat:
        $ref: '#/definitions/store.Seat'
      seat_id:
        type: integer
      session:
        $ref: '#/definitions/store.Session'
      session_id:
        type: integer
      user_id:
        type: integer
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for Tikceto
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Tikceto
paths:
...
  /seats:
    post:
      consumes:
      - application/json
      description: Creates a seat in a specified room
      parameters:
      - description: Seat payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.CreateSeatPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.Seat'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Creates a seat
      tags:
      - seats
  /seats/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a seat by ID
      parameters:
      - description: Seat ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Deletes a seat
      tags:
      - seats
    get:
      consumes:
      - application/json
      description: Fetches a seat by ID
      parameters:
      - description: Seat ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Seat'
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Fetches a seat
      tags:
      - seats
    patch:
      consumes:
      - application/json
      description: Updates a seat by ID
      parameters:
      - description: Seat ID
        in: path
        name: id
        required: true
        type: integer
      - description: Seat payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.UpdateSeatPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Seat'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Updates a seat
      tags:
      - seats
  /seats/session/{sessionID}:
    get:
      consumes:
      - application/json
      description: Fetches all seats for a specific session by session ID
      parameters:
      - description: Session ID
        in: path
        name: sessionID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of seats
          schema:
            items:
              $ref: '#/definitions/store.SeatWithMetadata'
            type: array
        "400":
          description: Invalid session ID
          schema: {}
        "404":
          description: No seats found for this session
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Fetch seats by session
      tags:
      - seats
...
  /tickets:
    post:
      consumes:
      - application/json
      description: Creates a ticket
      parameters:
      - description: Ticket payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.CreateTicketPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.Ticket'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Creates a ticket
      tags:
      - tickets
  /tickets/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a ticket by ID
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Deletes a ticket
      tags:
      - tickets
    get:
      consumes:
      - application/json
      description: Fetches a ticket by ID
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Ticket'
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Fetches a ticket
      tags:
      - tickets
    patch:
      consumes:
      - application/json
      description: Updates a ticket by ID
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: integer
      - description: Ticket payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.UpdateTicketPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Ticket'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Updates a ticket
      tags:
      - tickets
  /tickets/my:
    get:
      consumes:
      - application/json
      description: Fetches a ticket by ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Ticket'
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Fetches a ticket
      tags:
      - tickets
...
securityDefinitions:
  ApiKeyAuth:
    description: Use the Authorization header with a Bearer token to authenticate
      API requests.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"